+++ example_runner.py
@@ -64,6 +64,7 @@ import example.minibase.circle
 import example.minibase.circle_uneven
 import example.minibase.ellipse
 import example.minibase.rectangle
+import example.minibase.rectangle_uneven
 import example.minibase.slot
 import example.minibase.slot_uneven
 import example.minibase.hexagon